fix errors consoles
